新しいイラストの表示形式に対応するためモジュール分割
