+++ manifest.json
@@ -1,6 +1,6 @@
 {
 	"manifest_version": 2,
-	"version": "0.5.0",
+	"version": "0.5.1",
 	"name": "pixiv-download-image-with-name",
 	"description": "pixivの画像を、良い感じに名前を付けてダウンロードできるようにします。",
     "icons": {
